MultiLinkImpl
